requirements.txt
opencv-python>=4.11.0
mediapipe>=0.10.21
tensorflow>=2.19.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
